Fixes and optimizing, added 2 new block types

+++ include/client/gui/Inventory.h
@@ -16,6 +16,6 @@ static inline int Inventory_QuickSelectCalcWidth(int slots) {
 	return 54 + (slots - 2) * 20;
 }
 
-void Inventory_DrawQuickSelect(int x, int y, ItemStack* stacks, int count, int* selected);
+void Inventory_DrawQuickSelect(int x, int y, ItemStack* stacks, u8 count, u8* selected);
 
-int Inventory_Draw(int x, int y, int w, ItemStack* stacks, int count, int site);
+u8 Inventory_Draw(int x, int y, int w, ItemStack* stacks, u8 count, u8 site);
